bug fix, update not working as intended

+++ RoomDatabaseExample/app/src/main/java/de/test/roomdatabaseexample/AddEditNoteActivity.java
@@ -66,7 +66,7 @@ public class AddEditNoteActivity extends AppCompatActivity {
         data.putExtra(EXTRA_DESCRIPTION,description);
         data.putExtra(EXTRA_PRIORITY,priority);
 
-        int id = getIntent().getIntExtra(EXTRA_ID,-1);
+        long id = getIntent().getLongExtra(EXTRA_ID,-1);
         if(id != -1){
            data.putExtra(EXTRA_ID,id);
         }
